+++ src/doc/changelog.md
@@ -3,6 +3,8 @@
 - Updates the widget to `macversion` 2.
 - Updates the echarts library to 5.4.3.
 - Allows to give the input options as a JSON string.
+- Adds output endpoints for events of the chart (highlight, click, dblclick and hover).
+- Adds an option to allow merging data.
 
 ## 0.0.2 (2019-05-03)